fix warnings.
git-svn-id: svn://qnap.local/TwoTerm/branches/fix-gno-scrolling-region@3157 5590a31f-7b70-45f8-8c82-aa3a8e5f4507

+++ cpp/ring_buffer.h
@@ -56,7 +56,7 @@ public:
         _capacity = 0;
         // fill up the end, then wrap around to start.
         if (_ptr + dsize > Size) {
-            int amt = Size - _ptr;
+            size_t amt = Size - _ptr;
             memcpy(_buffer + _ptr, data, amt);
             _ptr = 0;
             dsize -= amt;
